v0.78.0
Fixes a few bugs including, I think, the pawn selector bug in the RimDialogue menu.
Adds as "Save" conversation feature.   This turns the selected conversation into a meme / cartoon thing that you can save as an image.
Added a discord button to the UI.
Adds compatibility for 1.6.  But Interaction Bubbles still does not have 1.6 compatibility so we a kinda stuck behind that for the moment.